+++ app/src/main/java/aefrh/es/aefrh/presentation/fiestas/details/FragmentFiestaDetails.kt
@@ -1,14 +1,9 @@
-package aefrh.es.aefrh.presentation.fiestas
+package aefrh.es.aefrh.presentation.fiestas.details
 
 import aefrh.es.aefrh.R
 import aefrh.es.aefrh.presentation.base.BaseFragment
-import android.os.Bundle
-import androidx.fragment.app.Fragment
-import android.view.LayoutInflater
 import android.view.View
-import android.view.ViewGroup
 import androidx.databinding.ViewDataBinding
-import com.custom.sliderimage.logic.SliderImage
 import kotlinx.android.synthetic.main.fragment_fiesta_details.*
 import timber.log.Timber
 
@@ -23,7 +18,10 @@ class FragmentFiestaDetails: BaseFragment() {
     override fun onViewsInitialized(binding: ViewDataBinding, view: View) {
 
         arguments?.let {
-            safeArgs = FragmentFiestaDetailsArgs.fromBundle(it)
+            safeArgs =
+                FragmentFiestaDetailsArgs.fromBundle(
+                    it
+                )
         }
 
         // add images URLs
