+++ app/src/main/java/aefrh/es/aefrh/presentation/fiestas/list/FiestasListAdapter.kt
@@ -1,4 +1,4 @@
-package aefrh.es.aefrh.presentation.fiestas
+package aefrh.es.aefrh.presentation.fiestas.list
 
 import aefrh.es.aefrh.R
 import aefrh.es.aefrh.databinding.FiestaItemBinding
@@ -12,7 +12,9 @@ import androidx.recyclerview.widget.DiffUtil
 import androidx.recyclerview.widget.ListAdapter
 import androidx.recyclerview.widget.RecyclerView
 
-class FiestasListAdapter: ListAdapter<Fiesta, FiestasListAdapter.ViewHolder>(VideoDiffCallback()) {
+class FiestasListAdapter: ListAdapter<Fiesta, FiestasListAdapter.ViewHolder>(
+    VideoDiffCallback()
+) {
 
     override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): ViewHolder {
         return ViewHolder(
@@ -34,7 +36,10 @@ class FiestasListAdapter: ListAdapter<Fiesta, FiestasListAdapter.ViewHolder>(Vid
 
     private fun createOnClickListener(epocaId: String): View.OnClickListener {
         return View.OnClickListener {
-            val directions = FragmentFiestasDirections.actionFragmentFiestaToFragmentFiestaDetails(epocaId)
+            val directions =
+                FragmentFiestasListDirections.actionFragmentFiestaListToFragmentFiestaDetails(
+                    epocaId
+                )
             it.findNavController().navigate(directions)
         }
     }
