+++ app/src/main/res/layout/fragment_fiesta_details.xml
@@ -2,7 +2,7 @@
 <layout
     xmlns:tools="http://schemas.android.com/tools"
     xmlns:android="http://schemas.android.com/apk/res/android"
-    tools:context=".presentation.fiestas.FragmentFiestaDetails">
+    tools:context=".presentation.fiestas.details.FragmentFiestaDetails">
 
     <data>
         <variable
