+++ app/src/main/res/layout/fragment_fiesta_list.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".presentation.fiestas.FragmentFiestas">
+    tools:context=".presentation.fiestas.list.FragmentFiestasList">
 
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
